requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
dj-database-url==3.1.2
Django==6.0.3
django-browser-reload==1.21.0
django-tailwind==4.4.2
gunicorn==25.1.0
idna==3.11
mercadopago==2.3.0
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
pytailwindcss==0.3.0
python-decouple==3.8
requests==2.33.1
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.12.0
reportlab==4.2.0
openpyxl==3.1.2
python-docx==1.1.2
